Fix bug when showing other content
In order to use smtweb to view content that was generated by somebody else one needs to set the datastore pk statically to 1.
This way it will always refer to the own datastore and load the content from a local folder that is synced between users (e.g. via dropbox).